enable REST response validation against OAS #16
Each JSON body in response is validated against OAS.
It fails with status code 500 if response is not valid.
Additional fields which are in JSON body but are not in OAS are
removed from the response.

Formatting of JSON body is also enabled when res.json() is called.

Depends on https://github.com/openBackhaul/ApplicationPattern/issues/120

Fixes #16

Signed-off-by: Martin Sunal <12826634+MartinSunal@users.noreply.github.com>